test: move utils to separate file, add tests for utils
