+++ package.json
@@ -39,7 +39,7 @@
     "pre-commit": "^1.2.2",
     "snazzy": "^9.0.0",
     "standard": "^16.0.1",
-    "tap": "^14.4.1",
+    "sinon": "^13.0.2",
     "tmp": "0.2.1"
   },
   "dependencies": {
