+++ v7.js
@@ -1,30 +1,6 @@
 const split = require('split2')
 const Parse = require('fast-json-parse')
-const pino = require('pino')
-const fs = require('fs')
-
-function getDestinationStream (destination) {
-  return (destination === ':2')
-    ? process.stderr
-    : fs.createWriteStream(destination, { flags: 'a' })
-}
-
-function getLevelNumber (level) {
-  if (typeof level === 'number' || !isNaN(Number(level))) {
-    return level
-  }
-
-  if (typeof level === 'string') {
-    const num = pino.levels.values[level]
-    if (typeof num === 'number' && isFinite(num)) {
-      return num
-    } else {
-      throw new Error('no such level')
-    }
-  }
-
-  throw new Error('no such level')
-}
+const { getDestinationStream, getLevelNumber } = require('./util')
 
 module.exports = (options) => {
   const filters = Object
